Create glossary.html
The [eSSIF-Lab Glossary](https://essif-lab.github.io/framework/docs/essifLab-glossary) as it stands today has been copied and turned into this HTML file for the purpose of enabling ToIP authors to use and/or refer to terms defined in/by eSSIF-Lab. This glossary will not be maintained. However, at some point in time, it will be automatically generated when necessary from the authoritative sources.